refactor: move logic for bulk select cancel button into approved tab
[#303 - Bulk reject Approved applications]
